add credentials directory
