+++ logtivly/log_hours.py
@@ -10,46 +10,12 @@ from oauth2client import tools
 from oauth2client.file import Storage
 import datetime
 import sys
+from credentials import credentials
 
 if __debug__:
     from workflow import Workflow, ICON_WEB, web
     from workflow.notify import notify
 
-
-
-# If modifying these scopes, delete your previously saved credentials
-# at ~/.credentials/sheets.googleapis.com-python-quickstart.json
-SCOPES = 'https://www.googleapis.com/auth/spreadsheets'
-CLIENT_SECRET_FILE = 'client_secret.json'
-APPLICATION_NAME = 'Google Sheets API Python Quickstart'
-
-
-def get_credentials():
-    """Gets valid user credentials from storage.
-
-    If nothing has been stored, or if the stored credentials are invalid,
-    the OAuth2 flow is completed to obtain the new credentials.
-
-    Returns:
-        Credentials, the obtained credential.
-    """
-    home_dir = os.path.expanduser('~')
-    credential_dir = os.path.join(home_dir, '.credentials')
-    if not os.path.exists(credential_dir):
-        os.makedirs(credential_dir)
-    credential_path = os.path.join(credential_dir,
-                                   'sheets.googleapis.com-python-quickstart.json')
-
-    store = Storage(credential_path)
-    credentials = store.get()
-    if not credentials or credentials.invalid:
-        flow = client.flow_from_clientsecrets(CLIENT_SECRET_FILE, SCOPES)
-        flow.user_agent = APPLICATION_NAME
-        if flags:
-            credentials = tools.run_flow(flow, store, flags)
-        else: # Needed only for compatibility with Python 2.6
-            credentials = tools.run(flow, store)
-    return credentials
 
 def get_spreadsheet_cell(service, spreadsheetId, projectStr="misc"):
     spreadsheet = service.spreadsheets().get(spreadsheetId=spreadsheetId).execute()
@@ -91,7 +57,7 @@ def get_sheet_title_and_column(service, spreadsheetId):
                 colNum +=1
 
 def get_projects_and_hours():
-    service, spreadsheetId = get_service_and_spreadsheetId()
+    service, spreadsheetId = credentials.get_service_and_spreadsheetId()
     sheetTitle, colNum = get_sheet_title_and_column(service, spreadsheetId)
     projectCells = 'B16:H19'
     initialProjectCellIndex = 16
@@ -101,7 +67,7 @@ def get_projects_and_hours():
     return colNum, sheetTitle, result.get('values',[])[0], result.get('values',[])[colNum+1]
 
 def get_project_cell(service, spreadsheetId, projectStr, sheetTitle, colNum):
-    service, spreadsheetId = get_service_and_spreadsheetId()
+    service, spreadsheetId = credentials.get_service_and_spreadsheetId()
     cols = ['c','d','e','f','g','h']
     columnLetter = cols[colNum]
     # it's not likely that there wil be more than 4 projects at a time
@@ -117,16 +83,8 @@ def get_project_cell(service, spreadsheetId, projectStr, sheetTitle, colNum):
     return '%s%s' % (columnLetter, initialProjectCellIndex + rowIndex), float(initialCellValue), projectNames[rowIndex]
 
 def py_main():
-    credentials = get_credentials()
-    http = credentials.authorize(httplib2.Http())
-    discoveryUrl = ('https://sheets.googleapis.com/$discovery/rest?'
-                    'version=v4')
-    service = discovery.build('sheets', 'v4', http=http,
-                              discoveryServiceUrl=discoveryUrl)
-
-    spreadsheetId = '1tRziPgOwgPBCYIQZuwa7Me1vk5_tfsAWb3mcpPICxEI'
-
-    projects, hours = get_projects_and_hours(service, spreadsheetId)
+    print(sys.path)
+    colNum, sheetTitle, projects, hours = get_projects_and_hours()
     index = 0
     items = []
     for project in projects:
@@ -139,19 +97,6 @@ def py_main():
 
     print(items)
 
-@staticmethod
-def get_service_and_spreadsheetId():
-    credentials = get_credentials()
-    http = credentials.authorize(httplib2.Http())
-    discoveryUrl = ('https://sheets.googleapis.com/$discovery/rest?'
-                    'version=v4')
-    service = discovery.build('sheets', 'v4', http=http,
-                              discoveryServiceUrl=discoveryUrl)
-
-    spreadsheetId = '1tRziPgOwgPBCYIQZuwa7Me1vk5_tfsAWb3mcpPICxEI'
-
-    return service, spreadsheetId
-
 def main(wf):
     """Shows basic usage of the Sheets API.
 
@@ -159,6 +104,7 @@ def main(wf):
     students in a sample spreadsheet:
     https://docs.google.com/spreadsheets/d/1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms/edit
     """
+
     if len(wf.args):
         sys.stderr.write("there ARE arguments!\n")
         sys.stderr.write(json.dumps(wf.args))
@@ -176,7 +122,7 @@ def main(wf):
             'values': values
         }
 
-        service, spreadsheetId =  get_service_and_spreadsheetId()
+        service, spreadsheetId =  credentials.get_service_and_spreadsheetId()
         service.spreadsheets().values().update(
             spreadsheetId=spreadsheetId, range=rangeName, body=body, valueInputOption="USER_ENTERED").execute()
