+++ logtivly/setup.sh
@@ -1,2 +1,2 @@
 pip install -r requirements.txt
-python make_credentials.py
+python credentials.py
